Fix build errors by wrapping useSearchParams in Suspense for auth pages
- Refactor `app/auth/login/page.tsx` and `app/auth/verify-email/page.tsx` to be Server Components.
- Extract client-side logic using `useSearchParams` to `components/auth/login-content.tsx` and `components/auth/verify-email-content.tsx`.
- Wrap client components in `Suspense` boundaries to satisfy Next.js static generation requirements.
- Add `components/auth/auth-page-loader.tsx` as a shared loading fallback.
- Verify build passes with `npm run build`.
