+++ components/auth/auth-page-loader.tsx
@@ -1,0 +1,9 @@
+import { Loader2 } from "lucide-react"
+
+export function AuthPageLoader() {
+    return (
+        <div className="min-h-screen flex items-center justify-center bg-background">
+            <Loader2 className="h-8 w-8 animate-spin text-primary" />
+        </div>
+    )
+}
